# perf(es/minifier): Make expression simplifier stateless and parallel