--- a/crates/swc_ecma_minifier/tests/terser/compress/global_defs/expanded/output.js
+++ b/crates/swc_ecma_minifier/tests/terser/compress/global_defs/expanded/output.js
@@ -3,7 +3,7 @@ function f(CONFIG) {
 }
 function g() {
     var CONFIG = {
-        VALUE: 1,
+        VALUE: 1
     };
     return CONFIG.VALUE;
 }