--- a/crates/swc_ecma_minifier/tests/terser/compress/global_defs/object/output.js
+++ b/crates/swc_ecma_minifier/tests/terser/compress/global_defs/object/output.js
@@ -2,7 +2,9 @@ function f(CONFIG) {
     return CONFIG.VALUE;
 }
 function g() {
-    var CONFIG = { VALUE: 1 };
+    var CONFIG = {
+        VALUE: 1
+    };
     return CONFIG.VALUE;
 }
 function h() {