--- a/crates/swc_ecma_minifier/tests/terser/compress/harmony/array_literal_with_spread_2a/output.js
+++ b/crates/swc_ecma_minifier/tests/terser/compress/harmony/array_literal_with_spread_2a/output.js
@@ -1,7 +1,19 @@
-console.log([10, 20, 30, 40, 50]["length"]);
+console.log([
+    10,
+    20,
+    30,
+    40,
+    50
+]["length"]);
 console.log(10);
 console.log(20);
 console.log(30);
 console.log(40);
 console.log(50);
-console.log(void 0);
+console.log([
+    10,
+    20,
+    30,
+    40,
+    50
+][5]);