--- a/crates/swc_ecma_minifier/tests/terser/compress/harmony/array_literal_with_spread_3a/output.js
+++ b/crates/swc_ecma_minifier/tests/terser/compress/harmony/array_literal_with_spread_3a/output.js
@@ -1,12 +1,24 @@
 console.log(10);
 console.log(20);
-console.log(void 0);
+console.log([
+    10,
+    20
+][2]);
 console.log(10);
 console.log(20);
-console.log(void 0);
+console.log([
+    10,
+    20
+][2]);
 console.log(10);
 console.log(20);
-console.log(void 0);
+console.log([
+    10,
+    20
+][2]);
 console.log(10);
 console.log(20);
-console.log(void 0);
+console.log([
+    10,
+    20
+][2]);