--- a/crates/swc_ecma_transforms_optimization/src/simplify/expr/mod.rs
+++ b/crates/swc_ecma_transforms_optimization/src/simplify/expr/mod.rs
@@ -1580,25 +1580,25 @@ impl VisitMut for SimplifyExpr {
 
     fn visit_mut_prop_or_spreads(&mut self, n: &mut Vec<PropOrSpread>) {
         self.maybe_par(cpu_count() * 8, n, |v, n| {
-            n.visit_mut_children_with(v);
+            n.visit_mut_with(v);
         });
     }
 
     fn visit_mut_expr_or_spreads(&mut self, n: &mut Vec<ExprOrSpread>) {
         self.maybe_par(cpu_count() * 8, n, |v, n| {
-            n.visit_mut_children_with(v);
+            n.visit_mut_with(v);
         });
     }
 
     fn visit_mut_opt_vec_expr_or_spreads(&mut self, n: &mut Vec<Option<ExprOrSpread>>) {
         self.maybe_par(cpu_count() * 8, n, |v, n| {
-            n.visit_mut_children_with(v);
+            n.visit_mut_with(v);
         });
     }
 
     fn visit_mut_exprs(&mut self, n: &mut Vec<Box<Expr>>) {
         self.maybe_par(cpu_count() * 8, n, |v, n| {
-            n.visit_mut_children_with(v);
+            n.visit_mut_with(v);
         });
     }
 }