--- a/crates/swc_ecma_transforms_optimization/src/simplify/expr/tests.rs
+++ b/crates/swc_ecma_transforms_optimization/src/simplify/expr/tests.rs
@@ -1544,24 +1544,3 @@ fn test_es6_features() {
         "function foo() {return `${false}`}",
     );
 }
-
-#[test]
-fn issue_1674() {
-    fold(
-        "
-            let foo = 'info';
-
-            var bar = [foo, (foo = 'other')][0];
-
-            console.log(foo);
-            console.log(bar);
-        ",
-        "
-            var _foo;
-            let foo = 'info';
-            var bar = (_foo = foo, foo = 'other', _foo);
-            console.log(foo);
-            console.log(bar);
-        ",
-    )
-}