--- a/crates/swc_ecma_transforms_optimization/src/simplify/expr/tests.rs
+++ b/crates/swc_ecma_transforms_optimization/src/simplify/expr/tests.rs
@@ -1539,8 +1539,8 @@ fn test_es6_features() {
         "class Foo {  constructor() {this.x = null <= null;} }",
         "class Foo {  constructor() {this.x = true;}}",
     );
-    // fold(
-    //     "function foo() {return `${false && y}`}",
-    //     "function foo() {return `${false}`}",
-    // );
+    fold(
+        "function foo() {return `${false && y}`}",
+        "function foo() {return `${false}`}",
+    );
 }